// src/client/src/constants.tsx
export enum DEVICE {
  DESKTOP = 1,
  MOBILE = 2,
}

export enum STATUS {
  PENDING = "PENDING",
  RUNNING = "RUNNING",
  DONE = "DONE",
  FAIL = "FAIL",
}

export enum COLOR {
  RED = "#EF5C4C",
  YELLOW = "#F3A631",
  GREEN = "#4BA15D",
}

export enum DATES {
  D1 = 1,
  W1 = 7,
  M1 = 30,
  M6 = 180,
  Y1 = 365,
  ALL = "ALL",
}